requirements.txt
zeep==3.4.0
xmltodict==0.12.0
requests-html==0.10.0
pyTelegramBotAPI==3.6.6